+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Changed:
 - Flexible figure size for some visualization
 - IO options validation and output directories creation logic
 - Device validate and use logic
+- dataset load logic
 
 Fixed: